Implement codec traits on `poplar` messages
Adds `Encode` and `Decode` implementations for the `poplar1` VDAF
protocol messages. The wire encoding of `poplar1` is not yet specified
by VDAF, so I have made some guesses here. In particular, I'm not
certain whether the length of some message members can be known
statically based on a VDAF instantiation and so have encoded them with
variable length in a few places.