Require codec traits on Vdaf associated types
Adds trait bounds into various associated types on trait `Vdaf` so that
we can (1) ensure that any VDAF implementation provides the necessary
`Encode` and `Decode` implementations on its associated types and (2)
ensure that clients of crate `prio` can work generically with traits
`Vdaf`, `Aggregator`, `Client` and `Collector` to encode and decode
messages. We also require `Clone + Debug` on a number of associated
types to save clients the trouble of specifying lots of trait bounds
when working with trait `Vdaf`.